Add files via upload
Created a tool to allow multiple .mblink playlists to be updated to the new format. I had near 100 playlists in Jellyfin at the time of migration so this script became a necessity for me. These playlists can reference multiple locations thanks to newpathlist, but this will need to be adjusted for each user's environment. This will walk through each file under rootdir and then examine if the file extension (fileext) is .mblink. If so, pathing within the .mblink file is changed from Windows style \\ to Posix / in temporary memory. This then prints off the results from memory, so these can be used under path_replacements and fs_path_replacements in the main script jellyfin_migrator.py. 

I considered adding this as a function in jellyfin_migrator.py but decided against it as I had to run the script a few times to get past issues with Unicode errors that arose from alternate languages. This granular control of getting the output from this script and then inserting it into the main script can help the user to detect if these errors exist, and see if they've properly set up their pathing in newpathlist.